Md Services

+++ messenger/src/main/java/app/domain/services/DeleteServiceDelivery.java
@@ -2,9 +2,9 @@ package app.domain.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import app.application.exceptions.BusinessException;
 import app.domain.model.ServiceDelivery;
+import app.domain.model.enums.Status;
 import app.domain.ports.ServiceDeliveryPort;
 
 @Service
@@ -19,11 +19,10 @@ public class DeleteServiceDelivery {
             throw new BusinessException("El servicio de entrega que intenta eliminar no existe.");
         }
 
-        // 2. Reglas de negocio opcionales (Integridad)
-        // Podríamos restringir la eliminación si el servicio ya está finalizado,
-        // pero dado que esta función suele ser administrativa para corrección de
-        // errores,
-        // permitimos la eliminación directa.
+        var status = service.getCurrentStatus();
+        if (status == Status.DELIVERED) {
+            throw new BusinessException("El servicio de entrega que intenta eliminar ya está finalizado.");
+        }
 
         serviceDeliveryPort.deleteById(id);
     }
